test/requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
collection==0.1.6
decorator==5.1.1
docopt==0.6.2
docutils==0.19
flex==6.14.1
genson==1.2.2
idna==3.4
jsonpath-ng==1.5.3
jsonpointer==2.3
jsonschema==4.17.3
pipreqs==0.4.11
ply==3.11
Pygments==2.14.0
pyrsistent==0.19.3
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
RESTinstance==1.3.0
rfc3987==1.3.8
robotframework==6.0.2
robotframework-jsonlibrary==0.5
robotframework-requests==0.9.4
six==1.16.0
strict-rfc3339==0.7
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validate-email==1.3
yarg==0.1.9
